+++ src/packageurl/contrib/url2purl.py
@@ -26,6 +26,7 @@
 
 import os
 import re
+import tarfile
 from urllib.parse import unquote_plus
 from urllib.parse import urlparse
 
@@ -615,3 +616,19 @@ hackage_pattern = (
 )
 
 register_pattern('hackage', hackage_pattern)
+
+
+@purl_router.route('https?://storage.googleapis.com/google-code-archive-downloads/v2/code.google.com/.*')
+def build_generic_google_code_archive_purl(uri):
+    # https://storage.googleapis.com/google-code-archive-downloads/v2/code.google.com/android-notifier/android-notifier-desktop-0.5.1-1.i386.rpm
+    _, remaining_uri = uri.split('https://storage.googleapis.com/google-code-archive-downloads/v2/code.google.com/') # android-notifier/android-notifier-desktop-0.5.1-1.i386.rpm
+    if remaining_uri:
+        split_remaining_uri = remaining_uri.split("/") # android-notifier, android-notifier-desktop-0.5.1-1.i386.rpm
+        if split_remaining_uri:
+            name = split_remaining_uri[0] # android-notifier
+            return PackageURL(
+                type='generic',
+                namespace='code.google.com',
+                name=name,
+                qualifiers={'download_url': uri}
+            )